# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IT Architect Assistant Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge/sources": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_knowledge_sources_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge Base Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceResponse"
                  },
                  "title": "Response List Sources Api V1 Knowledge Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Create Source",
        "operationId": "create_source_api_v1_knowledge_sources_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/sources/{source_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Source",
        "operationId": "get_source_api_v1_knowledge_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "knowledge"
        ],
        "summary": "Update Source",
        "operationId": "update_source_api_v1_knowledge_sources__source_id__patch",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/sources/{source_id}/archive": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Archive Source",
        "operationId": "archive_source_api_v1_knowledge_sources__source_id__archive_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/sources/{source_id}/disable": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Disable Source",
        "operationId": "disable_source_api_v1_knowledge_sources__source_id__disable_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/sources/{source_id}/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Source Documents",
        "operationId": "list_source_documents_api_v1_knowledge_sources__source_id__documents_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDocumentResponse"
                  },
                  "title": "Response List Source Documents Api V1 Knowledge Sources  Source Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Register Source Document",
        "operationId": "register_source_document_api_v1_knowledge_sources__source_id__documents_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceDocumentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/versions": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Versions",
        "operationId": "list_versions_api_v1_knowledge_versions_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge Base Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/app__schemas__knowledge__KnowledgeVersionResponse"
                  },
                  "title": "Response List Versions Api V1 Knowledge Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/versions/{knowledge_version_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Version",
        "operationId": "get_version_api_v1_knowledge_versions__knowledge_version_id__get",
        "parameters": [
          {
            "name": "knowledge_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__knowledge__KnowledgeVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/versions/{knowledge_version_id}/activate": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Activate Version",
        "operationId": "activate_version_api_v1_knowledge_versions__knowledge_version_id__activate_post",
        "parameters": [
          {
            "name": "knowledge_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeVersionActivateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__knowledge__KnowledgeVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Criticality": {
        "type": "string",
        "enum": [
          "required",
          "optional"
        ],
        "title": "Criticality"
      },
...
      "DocumentType": {
        "type": "string",
        "enum": [
          "normative",
          "architecture",
          "api",
          "technology",
          "other"
        ],
        "title": "DocumentType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeVersionActivateRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "KnowledgeVersionActivateRequest"
      },
...
      "SourceCreateRequest": {
        "properties": {
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Base Id"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "base_uri": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Uri"
          },
          "criticality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Criticality"
              },
              {
                "type": "null"
              }
            ]
          },
          "refresh_policy": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Policy"
          },
          "sync_mode": {
            "$ref": "#/components/schemas/SourceSyncMode",
            "default": "full_scan"
          },
          "source_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Metadata"
          },
          "active_flag": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Flag"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "name"
        ],
        "title": "SourceCreateRequest"
      },
      "SourceDocumentCreateRequest": {
        "properties": {
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "uri": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Uri"
          },
          "version_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Label"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "title",
          "uri"
        ],
        "title": "SourceDocumentCreateRequest"
      },
...
      "SourceResponse": {
        "properties": {
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "title": "Source Id"
          },
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "title": "Knowledge Base Id"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "base_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Uri"
          },
          "criticality": {
            "$ref": "#/components/schemas/Criticality"
          },
          "status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "refresh_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Policy"
          },
          "sync_mode": {
            "$ref": "#/components/schemas/SourceSyncMode"
          },
          "source_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_discovered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Discovered At"
          },
          "last_sync_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync Time"
          },
          "next_sync_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Sync Time"
          },
          "availability_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Status"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          },
          "latest_document_registered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Document Registered At"
          },
          "last_processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Processed At"
          },
          "last_processing_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Processing Status"
          },
          "last_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "last_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Message"
          },
          "update_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Run Id"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "knowledge_base_id",
          "source_type",
          "name",
          "base_uri",
          "criticality",
          "status",
          "refresh_policy",
          "sync_mode",
          "created_at"
        ],
        "title": "SourceResponse"
      },
...
      "SourceStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "unavailable",
          "disabled",
          "archived"
        ],
        "title": "SourceStatus"
      },
      "SourceSyncMode": {
        "type": "string",
        "enum": [
          "full_scan",
          "link_discovery"
        ],
        "title": "SourceSyncMode"
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "repository",
          "url_list",
          "manual_upload",
          "url",
          "local_folder",
          "catalog",
          "manual_registry"
        ],
        "title": "SourceType"
      },
      "SourceUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "base_uri": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Uri"
          },
          "criticality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Criticality"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "refresh_policy": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Policy"
          },
          "sync_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceSyncMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Metadata"
          },
          "active_flag": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Flag"
          }
        },
        "type": "object",
        "title": "SourceUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}